[FAB-6832] Add peer resource config bundlesource
This CR mimics the channel config bundle, and adds a peer resource
config bundle source which provides access to an underlying set of
channel config bundle and peer resource config bundle.

The underlying data structures are updateable via an atomic pointer
assignment, and stable views of the current configuration (totaling both
the channel configuration and the peer resource configuration) may be
obtained.

Change-Id: I84646c526b256aa140d23f4d7982c91dd67b0aa3
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>