[FAB-6696] Fix printOrg output to deeply marshal
The configtxgen -printOrg function currently prints a version of the
organization definition with the MSP definition opaque.  This is a
problem for the configtxlator APIs which expect for the JSON to be
deeply marshaled.

This CR simply fixes the printing by changing the type passed in to be
the proper annotated dynamic type, rather than the base ConfigGroup
type.

Change-Id: I7952ad7c052f6f063ab747f323c2ab531340887f
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>