Retire dormant maintainers
I'd like to propose that we retire three of the Fabric maintainers
who have not been active as maintainers for the past six months or more.
Tamas Blummer (DAH), Gabor Hoszzu (DAH), Sheehan Anderson (salesforce).

To quote from the readthedocs maintainers policy:

"Maintainers may be removed by explicit resignation, for prolonged
inactivity (3 or more months), or for some infraction of the code
of conduct or by consistently demonstrating poor judgement. A
maintainer removed for inactivity should be restored following
a sustained resumption of contributions and reviews (a month or
more) demonstrating a renewed commitment to the project."

Added a 'Retired Maintainer' section.

Change-Id: I2746ac3d4234da3d054b410562633b8e2b3b6323
Signed-off-by: Christopher Ferris <chrisfer@us.ibm.com>
Signed-off-by: David Enyeart <enyeart@us.ibm.com>