[FAB-6723] chaincode lifecycle for peer services
The gossip module needs to be updated when a chaincode that is installed,
is deployed on the channel - or installed on the peer if it is already
deployed on the channel.

This change set adds a chaincode lifecycle object which tracks
of the deployed chaincodes that are installed locally on the peer.

It attaches to the join channel events and registers subscriptions
for each channel.
The subscriptions are proxies that register themselves to the ledger
ChaincodeLifecycleEventListener, and these updates
are forwarded to the lifecycle itself.

The lifecycle updates gossip (with the chaincodes that are
deployed on the peer) via a listener that is created
at peer startup.

So, the flow is:

At peer startup:

lifeCycle.AddListener(updateGossipWithChaincodes)

for each channel at join channel:
  sub <--- lifeCycle.NewSubscription(channel)
  ledger.chaincodeEventManagement.Register(sub, channel)

So, the ledger.chaincodeEventManagement pass updates (when a chaincode
that is installed on the peer is also deployed, or when a deployed
chaincode is installed on the peer) to the corresponding subscription
and it forwards to the LifeCycle which forwards to gossip.

This is needed because the ledger events infrastructure is:
1) Too deep inside the ledger code and isn't easily accessible
   to the various peer services such as service discovery and gossip
2) Doesn't enumerate installed and instantiated chaincodes at startup,
   but does it only for new runtime events. However, a peer needs
   to start announcing the chaincodes it endorses right away.

Later on, the discovery module could query cached metadata
for the policies, in order to compute endorsers for chaincodes.

For that we would need to link this LifeCycle to a component that retreives
LSCC metadata from the ledger, in a similar manner which is done now in this existing
code, but for chaincodes that aren't installed on the peer.
For now, we only cache the metadata of the installed+deployed chaincodes.
This is because the ledger chaincodeEventManagement gives us updates,
however we need to pass to gossip a full snapshot each time.

The change set comes with unit tests that provide 100% code coverage in
ginkgo flavor and served with a side dish of autogenerated mocks.

Change-Id: I62020e592291f50dbfd2fe644acf2cbdde84ce62
Signed-off-by: yacovm <yacovm@il.ibm.com>