[FAB-6746] Fix modification of root config group
A recent change was made to the configtx processing which causes a panic
when the path of a config element is empty.  This case can never happen
for any config element, except for the very root config element.  This
CR modifies the code to simply not traverse the policy manager selection
code if the path is empty, as this indicates the element is part of the
root policy manager.

Change-Id: Ibc73ad155a51b79051c83a84cd8d615d432506e2
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>