[FAB-6741] Deprecate stateInfo.Metadata
The stateInfo.Metadata field was used in v1.0 for storing the peer's
ledger height.

In v1.1 we added a properties proto structure field that would replace
the metadata for ledger height.
In v1.2 (next release) we deprecate the metadata and as a result, peers
in v1.0 won't be able to gossip in channel context with peers
from v1.2

Change-Id: Ieefa98563500760eb23bf579f93ed786b935acf4
Signed-off-by: yacovm <yacovm@il.ibm.com>