[FAB-5921] add channel to proposal contexts
Proposal contexts are currently keyed of just transaction IDs.
This prevents CC-2-CC calls where the called CC is on a different
channel (we want to use the same TXID for the entire CC call chain).

This CR adds Channel ID to the context key. In addition to fixing the
above bug, this also has a couple of good side effects
    . Chaincode can now get the ChannelID of proposals (they could do
      it from the proposal header but that would not work when calling
      a CC on a different channel)
    . proposals with the SAME transaction id can now be sent to the
      SAME chaincode on multiple channels (this is different from the
      CC-2-CC issue the bug was raised on)

. patch-2 removed a check that should never have been there (forgot to
  remove it in previous checkin). Also found a bug that prevented that
  from kicking in.
. patch-4 rebuilt protos to pick up protos/common/collection.pb.go
  (hopefully will fix z built UT failures in protos/utils)
. patch-5 rebase and fix some "go vet" issues now that we are using go vet

Change-Id: I7e4333d4219a0927885ed787fda05e0dcaab25fd
Signed-off-by: Srinivasan Muralidharan <srinivasan.muralidharan99@gmail.com>