[FAB-6788] Mutual TLS client connections for peer
Most of the plumbing to enable the peer to make
mutual TLS connections to other peers and to the
ordering service has been in place for a while
but some minor refactoring was required in order
to add support for using mutual TLS with the
ordering service as well as with peers.

comm.CASupport is also now being used in the
orderer so it no longer made sense to add
specific function(s) to it in support of
client connections which will be made by the
peer.  So a new type comm.CredentialSupport
which "extends" CASupport was added.
CredentialSupport includes a ClientCert field
and the GetPeer/DeliverServiceCredentials functions
were moves to this object.

Change-Id: I0abf423a3c48d29408460ca2fe54cf14e7d6039c
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>