[FAB-6855] Add CouchDB index warming
CouchDB does not automatically update indexes when records are
inserted or updated.  The update is performed for the index
during the query operation.   This can cause delays during
the query.   Subsequent queries then take advantage of the updated
index.

This improvement will add an "index warming" step to refresh indexes
following an insert or update.

The addition is to the EnsureFullCommit step.  This will execute an
index warming action for each batch.  The index refresh will need to
be called for each index defined.

Change-Id: I4de7f8049be8d5ff092c84a42d250ab7c361c492
Signed-off-by: Chris Elder <chris.elder@us.ibm.com>