[FAB-6840] Consolidate configtxapi to configtx
Once upon a time, the configtx package did far too much.  It exposed the
channel configuration for each component, it handled transactional state
for the policy manager, and the other transaction components, and
generally became something that too many packages linked to.  Because of
all the different packages which linked to it, the interface definitions
were factored out into their own package configtxapi.

Since then, the vast majority of this function has been factored out
into other more appropriate places, and there are no longer dependency
cycles.

This CR moves the interface definitions back into the package, and
changes the rather generic 'configtx.Manager' name to
'configtx.Validator' (as its only purpose now is to validate configtx
updates before generating a new configuration).  This renaming ripples
throughout a fair bit of the code, but should be reasonably easy to
review.

Change-Id: Icd8494fc727a2474fc28f15945da7d0548f49f23
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>