[FAB-6854] Point to proper version of samples
fabric-samples has two active branches:
- release
- master

The content in release branch is based on v1.0.x
releases.  The release branch is also the default
branch when doing a git clone.

The master branch of fabric-samples has been
updated for v1.1.x so we need to make sure that
the latest/master docs instruct people to clone
the master branch of fabric-samples

Change-Id: I169e4591b982ad6249c23c4793ddc9da6014b6c9
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>