[FAB-5222] Correct misleading CLI output on join
The peer join command submits a join proposal request asynchronously, so
it does not know the outcome of the request immediately after command
is finished. But the output it prints is "Peer joined the channel", which
is misleading. The change set corrects this message to "Successfully
submitted proposal to join the channel"

Change-Id: I3e468547d128c0118769d5b551c465838567940b
Signed-off-by: Anil Ambati <aambati@us.ibm.com>