[FAB-6333] Add chaincodeAddress to peer
As described in issue FAB-6333, when using hostname for fabric peers,
the changed IP address after reconnect causes a problem as the
original chaincode listener continues to listen on the old address
that was resolved by the hostname when the peer was started.

The CR will fix this issue by adding the param "chaincodeAddress"
to peer, which is the connecting host name for the chaincode, while
the param "chaincodeListenAddress" is stil used to set listen address.
Also, it updates core.yaml to reflect this change.

At last, it refactors getting chaincode endpoint part into a single function.

What's more, it removes unused "peerAddress" from core.yaml chaincode section.

Change-Id: I85e273a10b66859077f10c7b0774cda55d3e9cb8
Signed-off-by: Boliang Chen <cblsjtu@gmail.com>