[FAB-6777] Incorrect comment about cc events
This CR updates the comment for SetEvent to ensure it accurately
reflects the expected result of setting a chaincode event. It previously
stated that a chaincode event would only be sent if the transaction
was validated, which is incorrect. The chaincode event will always be
available in the block event whether the transaction was validated or
not because these events are stored in the committed block.

Change-Id: I0eabc67184b9ef2a4cab8901ee2fd9f0ee2e9d5b
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>