[FAB-6803] Rename capabilities key V1.1 to V1_1
We are using an old version of Viper, which has problems locating
keys with period. It uses period as delimiter for the path of a
nested structure, therefore, foo.bar.V1.1 would result in
foo:
  bar:
    V1:
      1
instead of
foo:
  bar:
    V1.1:
as we expected. This patch simply renames "V1.1" to "V1_1" to avoid
this problem.

Change-Id: Iebfa5ca40038beacd2bf4ccff3639b9e5e4ce0f0
Signed-off-by: Jay Guo <guojiannan1101@gmail.com>