[FAB-6927] Generate TLS client certs for users
Now tha that both the peer and the orderer
support mutual TLS, cryptogen needs to
generate TLS client certificates for the
users it generates.

All of the material was already being generated,
but all TLS certs were name server.* and it is
more appropriate / more clear to name TLS
certs intended for client usage client.*.
So internally a new parameter is added
which specifies the type of node and maps
node type to client or server as appropriate.

Change-Id: I510a07335f4c685367ff941ab6c63a0203a04bd1
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>