[FAB-6227] custom processor for resource configs
This CR introduces a custom transaction processor that processes
resource config update transactions. Ledger invokes this processor
for transactions of type CONFIG, and PEER_RESOURCE_UPDATE
during the validation and commit phase of a block processing.

For CONFIG, this simply stores the config in the statedb.
For PEER_RESOURCE_UPDATE - In a normal course, this validates the transaction
against the current resource bundle, computes the full configuration, and
stores the full configuration if the transaction is found valid.

However, during initialization of ledger (i.e., either the ledger is being
created from the genesis block or the ledger is synching the state with the
blockchain during start up - possibly prommpted by a peer crash last time),
the full config is computed using the most recent configs from statedb, if found.

Change-Id: Ia18432e8cede18d378b820c13019acc2995197a4
Signed-off-by: manish <manish.sethi@gmail.com>