[FAB-7224] Enhance custom tx processor
This CR enhances custom transaction processor in ledger such that
the processor is able to distinguish whether it is being invoked
during ledger initialization or for on-line tx processing.

In the former situation, the processor needs to just rely on the
states stored previously in the ledger and in the latter it can assume that
the peer is up and running and hence other peer structures can be used.

Other than this, in the initialization situation only valid transactions are
passed to the processor which makes the custom validation in the processor
optional.

Change-Id: Idff54ab124dd588dfa5172a5b882e521217ca62a
Signed-off-by: manish <manish.sethi@gmail.com>