[FAB-2641] Prevent loop while gossiping msgs
Currently while gossip forward the message during distribution it
selects peers from the current membership view, while not filtering
the peer which delivered the message. As a consequence message might be
send to the source peer one again. This commit extends the peers
selection logic by adding the check which allows to avoid sending same
message to the source peer back again.

Change-Id: I76ceafba2335c27e1186f511d3cdab8a24d600af
Signed-off-by: Artem Barger <bartem@il.ibm.com>