[FAB-7066] Modifying the enccc example chaincode
The enccc_example.go was modified to clearly show the user
the crypto function(s) being available for transactional data
that has transient data passed for operations to include:
- encrypt
- decrypt
- sign
- verify

The patch set removed the 'PUT' and 'GET' parameters for sample clarity.
The README.md was also modified and the go tests modfied to
test the modified and new functions added.

Change-Id: I16b7894c1118cc7b9f722801ffe3a113be5dbece
Signed-off-by: Barry Mosakowski <barry_moz@yahoo.com>