[FAB-7100] Keepalive options not set for shim
The gRPC keepalive refactoring left out
the configuration of the keepalive
options for the shim to use when
connecting to the chaincode server.

This change rectifies that as well as
simplifying the code by removing the
specific chaincode client connection
functions.

Please note that the keepalive settings
for chaincode are statically configured.

Change-Id: I80aa26a4e69b7d85c5d4b32a53bbe37ed70f17df
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>