[FAB-7237] Specify TLS client key pair for peer
Prior to this change, the peer used the
TLS server key pair when making client
TLS connections.  In many cases, this
will be fine, but in cases where the
TLS server key pair is issued by
a public CA (e.g. Verisign), this is
an issue as now any client certificate
issued by Versign will be valid.

So this change simply adds an optional
TLS client key pair to the peer config.

Change-Id: I0195407ef0ecef58059e726a6c9eb6e473926cc8
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>