[FAB-7581] Enabling couchdb index creation
This CR introduces
1) 'ChaincodeLifecycleEventListener' - this needs to be implemented
by couchdb implementation in order to get a chance to create indexes
upon chaincode install or deploy. This function will be invoked for a
chain + chaincode combination only if the latest chaincode event
(i.e., deploy on a chain or install on peer) makes the state such that
the chaincode is deployed on the chain and is installed on the peer.

2) Chaincode event management module - this exposes two functions that
are to be invoked from chaincode deploy and chaincode install path
respectively. Internally, it synchronizes the channel creation,
chaincode install, and chaincode deploy so that we do not miss events
for index creation

3) Functions for getting the status of a chaincode (deployed/installed).
Further, if installed - extract the statedb artifacts from the chaincode
package. These functions are used by the module listed in (2) above

4) statelistener functionality - This allows for registering an arbitrary
listener for state changes over a namesapce. This is used for listening
the changes over 'lscc' namespace that indicate the deployment of one or more
chaincodes in a block

5) Enabling couchdb index creation upon chaincode deploy using the all above

6) In a separate CR, one function (namely, 'HandleChaincodeInstall') needs
to be invoked from chaincode install code path so that index creation is
enabled during chaincode install as well

Change-Id: I928f6ffd10f7985cbef63a70f76da920a4ca4c56
Signed-off-by: manish <manish.sethi@gmail.com>