[FAB-7497] retrieve cc data from LSCC
As per the original design of FAB-6042, the main repository for chaincode
data was to be the resource tree. As a consequence, code was introduced to
retrieve chaincode data before execution of a chaincode invocation (or of a
cc2cc) from either LSCC or from the resource tree (depending on the capability
of the channel). According to FAB-7497 however, chaincode data is always
stored in LSCC (initialized with data in the config tree at deploy/upgrade
time). This change set therefore removes the alternate code path.

Change-Id: Ie4757d6c365e7c381cd4bc169a41c509066fc4da
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>