[FAB-7434] Normalize UID/GID in cc packages
When creating a chaincode package, if
one uses the same code but it happens
to be owned by a different user/group,
a different package is produced and this
can result in a fingerprint mismatch
at instantiate time.

While people should actually install the
same package and not individually create
their own packages, creating packages
given identical inputs/code should be
repeatable on different machines so
this CR simply normalizes the UID/GID
when adding files to the archive.

Change-Id: I04b6ddde00bc80d98cf55de66ae9ff005f6b2aa7
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>