[FAB-7450] Generalize TLS cert extraction from message
FAB-6868 added a utility method that extracts the TLS cert hash from the
channel header of an envelope, and then compares it to the TLS cert hash
of a gRPC context.

Since we need to do that for the events
(which don't use the same envelopes),
it makes no sense to implement this again for
a new protobuf message type.

What we should ideally have is that utility to accept a function that
extracts the TLS cert hash from the protobuf message,
and then it can be reused for all our protocol messages.

This change set changes the utility method and adds that function.

Tests were modified accordingly.

Change-Id: Ia6c7bfa98490dedb71279239b4a3d6111a29f6ba
Signed-off-by: yacovm <yacovm@il.ibm.com>