[FAB-7491] client TLS cert support for gossip
This change set adds client TLS certificate support for gossip.
The handshake method now selects the certificate to hash and attach
its binding to the handshake message according to whether the peer
initiated the connection or not.

The change set adds the support in the form of a struct that
wraps the TLS certificates with atomic references.

This is in order to prepare for the future where we might have
dynamic TLS certificate updates.

Unit tests haven't been added, because I changed the test code
to always have a different client and server certificate,
and in case only the server certificate is present than the code
that is executed in the production path is code that is already
tested, because it is computed by code in core/peer/start.go

Change-Id: I1edddb2321c629f88080510befe1db26fa0b6925
Signed-off-by: yacovm <yacovm@il.ibm.com>