[FAB-7497] validate txes against LSCC cc data
As per the original design of FAB-6042, the main repository for chaincode
data was to be the resource tree. As a consequence, changes to both VSCC
and the validator code were introduced to retrieve chaincode data from
the resource tree and to use a policy pointer vs. policy literal in VSCC.
According to FAB-7497 however, chaincode data is always
stored in LSCC. This change set therefore removes the alternate code path.

Change-Id: I02421ac6c607472968e71f5969daa0a3bd7d2b07
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>