[FAB-7552] IdentityInfo in gossip identityStore
The discovery service needs to output identities of known peers.

Gossip stores the identities in an identity mapper,
which could be extended to return a snapshot of its identities.

This change set makes it so that the identity mapper also saves
the organizations of peers upon their insertion, and adds
the needed method.

Change-Id: I94b032781bae5d07a3d9fef586273ac8bfcc2e4d
Signed-off-by: yacovm <yacovm@il.ibm.com>