[FAB-7569] Fix failures in CouchDB unit tests
TestIndexOperations creates and drops indexes
in CouchDB and then queries the list of indexes.
Since the list of indexes is updated async in CouchDB,
sometimes the list of indexes is not as expected,
causing TestIndexOperations to fail.

This change adds a 100ms delay before the index
list is retrieved in the unit tests.

Change-Id: I87debd03567c4f1bc3f4d6366f28ed2c897f14a8
Signed-off-by: David Enyeart <enyeart@us.ibm.com>