[FAB-6938] Cannot create CouchDB database for channel
In rocket.chat there was a post about an error:
when i am running join peer request some times it's giving
Error:file_exists, Status Code:412, Reason:The database could
not be created, the file already exists

I believe this can happen if peer is retrying a create database
request against CouchDB.  A prior attempt to create database came
back as a timeout failure to peer, but in the meantime the database
got created in CouchDB.  The next retry would get the error above.
This scenario should be handled in peer code.

Change-Id: I01ecd3a5fde8a9042f6a18181376dec267d0cc15
Signed-off-by: Chris Elder <chris.elder@us.ibm.com>