[FAB-7633] Cleanup dead comm code
With FAB-7490, the peer CLI communication code
was refactored to use the comm.GRPCClient.

There are a few functions which are no longer
called in core/peer which have now been
removed.

Additionally, the comm.InitTLSForPeer function
is no longer needed by any of the main code paths.
As it was only included in a few tests (but never
actually executed) , it has been removed as well.

An additional benefit is that this removes viper
from the core/comm package as well.

Change-Id: I0c25f9896dc7caf8cb8d86d6f85ce89270c34b00
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>