[FAB-7641] Fix panic when iterating over private data
The gossip pull delegates fetching data from the data retriever which is
a bridge between gossip and the transient store.
The data retriever quries the ledger and iterates over an iterator until
Next() returns nil (or error).
However, Next() can also return an object that has its inner fields be nil
which represent that this current data element has no data in it, and
iteration must continue on.

This change set makes the code skip if the inner element is nil,
and adds a unit test for this.

Change-Id: Ica249e3b3a5081b1fc7357baa45bf9d0bc209e11
Signed-off-by: yacovm <yacovm@il.ibm.com>