[FAB-7696] Add couch indexes to node cc install pkg
This change packages index metadata for node.js
chaincode. It has the same behavior as the golang
packaging recently introduced in FAB-7132.

Allow indexes that support chaincode queries to be defined
alongside chaincode.  Any index defined in chaincode directory
META-INF/statedb/couchdb/indexes will get added to the chaincode
deployment spec code package, so that when the chaincode
is both installed on a peer and instantiated on a channel,
the indexes can be automatically deployed to couchdb.

Change-Id: I1670e77508f496fd839a4df2480350184c870fab
Signed-off-by: David Enyeart <enyeart@us.ibm.com>