[FAB-7808] Fix inconsistent configtx.yaml
The sample and example configtx.yaml file have profiles with similar
names, but slightly different content.  As this is a sample, it should
be fixed to read consistently.  This ensures that the capabilities
section for the channel is always specified along with the one for
capabilities in the orderer.  It also renames the alias
"GlobalCapabilities" to "ChannelCapabilities" as this is how the doc
refers to them.

Change-Id: I407918c604de1ee546d101edccdcb32693cdf043
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>