[FAB-7399] Check for uninitialize config state
The v1.1 peer code added config transaction processing to store the
channel config in the state database.  This is the more obviously
correct way to store config, but on upgrade, the state database will not
already have stored a copy of the config.

This CR adds back the old config processing code to extract the channel
config from the config block in the event that it is not stored in the
statedb.

Change-Id: Ied88406d92dd8d6d33d45a6d5eb3b05375270905
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>