[FAB-8147] Doc change 'update' to 'upgrade'
The documentation around upgrade frequently uses the word 'update' as a
synonym for 'upgrade'.  Since channel configuration is 'updated', not
necessarily as part of 'upgrade', this can be confusing for users.

This CR replaces the word 'update' with 'upgrade' where appropriate, and
tries to use other words like 'replace' in places where upgrade is
ambiguous.

Change-Id: I2000f9ce00a46314f0b28e9b79e5ecadb7162b78
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>