[FAB-8044] Add config option for defer index warming
CouchDB index warming is executed per block, and as a result it
may become an unexpected resource hog in some scenarios, such
as when a high rate of updates going on with very small Batchsize.

A new config option warmIndexesAfterNIntervals has been added to
prevent index warming from happening on every block.  Default is
configured as 1.

Change-Id: If7198a4140e27d312ee7caa2f667bf2c80df3657
Signed-off-by: Chris Elder <chris.elder@us.ibm.com>