[FAB-7952] Improve unclear and generic error message
This CR changes the error message reported when a proposal is processed
for a channel that either does not exist or the creator is not a member
of. The message is still left vague ("access denied") to avoid potential
malicious users scanning for channels but echoes back the channel and
and org name supplied by the user to provide some troubleshooting help
for cases where there was a typo in the channel name or the user had an
unexpected MSP configured. The originating error message is logged as-is
on the peer to allow peer admins to know the exact root cause of the
error. This CR also updates the errors throughout the file to use the
github.com/pkg/errors package.

Change-Id: Id3589ceb56fa817a3b95026bb3e3e34629346f80
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>