[FAB-8277] Make channel templates predictable
Presently, before applying a channel creation transaction, a 'template
config' is created by copying various pieces of the orderer system
channel into a new config, then applying the channel creation tx as an
update.

This works well in greenfield deployments, but as the orderer system
channel mutates, so does the template config.  Users would have to
deduce the correct versions to use in computing their config updates,
which becomes onerous and requires out of band communication.

There is no obvious benefit to preserving the versions from the orderer
system channel, so this CR predictably sets them to zero.

Change-Id: I11e379104dfd708eb7d31d3f75d516e0e981d39c
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>