[FAB-8182] Remove old configtxgen doc
The new command doc layout makes the old configtxgen document
superfluous.  However, there was still some content in the old doc that
was useful to preserve, and has not been moved to the command doc.

This CR also updates the referenced profiles to be the more recent V1_1
capability profiles, as was suggested in a review of the doc, but
deferred for a later CR.

Change-Id: I77a9e8e56c8fbf19e27b040279a4b74150f17ffb
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>