[FAB-8570] Reduce log severity if existing block added
Gossip enqueues blocks into the payload buffer via different mechanisms:

    State transfer anti-entropy
    Pull
    Push

Whenever an attempt to add a block to the payload buffer which was
already added before by some other pathway - a warning is logged.

I think we should make that warning a debug, since it's not something
that users should be worried about.

Change-Id: Ia735473e33ddff3f8787c121f39a6d15e54baa04
Signed-off-by: yacovm <yacovm@il.ibm.com>