[FAB-8923] remove utxo example
The example does not compile without additional external dependencies
and when we attempted to follow the instructions in the README.md to
create the docker image, it failed.

If the sample is still needed, development effort is required to bring
it back to a usable state. Since it's unusable in its current form,
removing from the tree to avoid additional confusion.

Change-Id: Ic0261ec11ce95319586f11e75f65ca68eea16f04
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>