[FAB-9002] Define new application v1.2 capability
For all the non-backwards compatible work, including the new chaincode
lifecycle and resources config changes, we must use a new capability to
determine whether the behavior is enabled or not.

This CR adds one for the application portion of the config.  At the time
of this writing there are no features which would introduce
incompatibilities at the orderer or channel level, so those capabilities
are being left in place.

This CR also removes the assorted 'V11' profiles, as maintaining a set
of profiles for each version is unmaintainable and confusing.  For this
and subsequent releases, configtx.yaml will contain only defaults for
the current version of fabric.  Users may of course override these
defaults by modifying the config file.

Change-Id: I08729fef76292044440ebc7e4f39145b0640afc4
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>