[FAB-9003] Add chaincode lifecycle capability
Although chaincode lifecycle is projected to be part of Fabric v1.2, for
development purposes, tying it to the V1_2 capability upfront is not an
option, as until everything is in place, we need to keep the code path
disabled.  Therefore, this CR defines a new capability for chaincode
lifecycle which may be turned on separately.  It is marked as
experimental in the spirit of being able to cut a release at any point.
However, the expectation is that this capability will disappear and the
chaincode lifecycle will be enabled through the v1.2 capability.

Change-Id: I61cdd106c907b4676d35eb1f288f69e578c83aad
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>