[FAB-9059] add metadata to chaincode platforms
Metadata was introduced into the chaincode package for
handling couchdb indexes. However the implementation
assumed chaincode will be in the targz format inside
ChaincodeDeploymentSpec. The CAR platform does not use
targz format and this exposes the weakness with the
implementation - the metadata should really be derived
from the platform interface.

This is the first of the series that moves chaincode
metadata to the platform interface. This only lays the
platform foundation (via a "MetadataProvider" supplied
by each platform) and does not have the rest of the code
actually use it. That will be done in the next CR.

Change-Id: I7a121cda03a7da543da5342a2f692195bebeb45c
Signed-off-by: Srinivasan Muralidharan <srinivasan.muralidharan99@gmail.com>