[FAB-9203] Maintain versions of collection configs
This CR introduces the code for maintaining different
versions of collection configuration. The primary usage
is for gossip to use this information during pulling
and disseminating the private data from committed data.

The eligibilty of the recieving peer for a data committed
at height 'H' is expected to be determined by retrieving
the most recent collection configuration below height H
(using the api 'MostRecentCollectionConfigBelow...' introduced
in this CR)

Change-Id: I14c4dc5fdfa97e747dc3ef94989d355467329530
Signed-off-by: manish <manish.sethi@gmail.com>