[FAB-9479] Filter out MSP IDs before computing layouts
This change set makes the endorsement analysis logic to filter
out principal sets that have MSP IDs which don't have peers
with the chaincode installed at all.

This is needed for implementing support for cc2cc calls,
because when computing cc2cc endorsement policies we compare
principal sets, and merge principal sets such that one set
contains the other.

However- even though a policy might have an MSP ID in it,
doesn't mean the peers of that MSP ID have the chaincode installed on them.

Therefore, we need to prune these principal sets beforehand in order
for them no to appear as input to the combined endorsement policy evaluation
which would be implemented in future change sets.

The change set also prepares the ground for cc2cc and collection
support by changing the signature of PeersForEndorsement to have
a chaincode interest which aggregate several chaincodes and collections.

Change-Id: If382b4254797684ffef20e182ae744f5c5f3dfa7
Signed-off-by: yacovm <yacovm@il.ibm.com>