[FAB-9516] Split cclifecycle update to 2 phases
The lifecycle object that caches the deployed chaincodes of the peer
receives a notification from the ledger before the install of the
chaincode actually happens.
The install might fail, and then gossip would publish false information.

For this - the ledger listener separated the notification into 2 steps:

 1.  Notify about the deployment, which happens just before the install
 2.  Notify whether the deployment was successful or not, which is only
   relevant for install.

This change set moves the state update of the lifecycle object to the DeployDone()
phase (the 2nd phase from above)

Change-Id: I495f73278ce4e7d017085a6df4043794bc7005a7
Signed-off-by: yacovm <yacovm@il.ibm.com>