[FAB-9580] Create SCC instances by constructor
This is the first in a series of CRs whose goal is to restructure the
existing chaincodes under the principles of depencency inversion.

Today, the assorted SCCs call out into the "system chaincode provider"
or other singletons which makes mocking difficult and initialization
haphazard.

This CR is a first step which pulls the singleton references out of the
Init functions and pushes them into constructor functions.  The next
step will be to modify these constructors such that they accept their
required dependencies.

Change-Id: I8ee1343ed8acad12e4445f90c73a08c7fd0182fe
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>