FAB-9568 Remove package scoped syscc functions
The assorted DeploySysCC and RegisterSysCC functions are all package
scoped and imply some operations on some local data.  Additionally, as a
bit of a hack, the chaincodes slice was modified to not actually
initialize the chaincode, but instead defer initialization through a
function.

This CR fixes up both of these issues, creating a true 'Provider' for
system chaincode with instance scoped verbs.

Change-Id: I499bb1bdc8469302626d39c8a0bb59ea2da2757d
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>