[FAB-9677] Restructure state DB encoder
In the current implementation, levelDB encodes the KVS version as part of
the KVS value, whereas couchDB encodes the KVS version as an additional
map entry. The code to package version and value into a byte array therefore
only belongs to the levelDB code, and not the generic stateDB code.

Change-Id: I673ff84d95c92b1829e9a0f37b536deb78314242
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>
Signed-off-by: Matthias Neugschwandtner <eug@zurich.ibm.com>