[FAB-9688] Remove dangerous unused CCID.ChainID
The CCID struct has a ChainID field.  This field affects the output of
GetName().  There is a comment in GetName() indicating that if unset,
the chaincode must be a chainless system chaincode, however, this field
is in fact never set.  If it were, the rest of the chaincode container
logic would break, and we would spawn a container per chaincode per
channel, instead of just per chaincode.

This CR removes this unused field and the logic in GetName around it.

Change-Id: I9f666d7f2c2942372a363bc301f275c56fe55228
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>