FAB-9762 Remove VMCResp structure
The response portion of the VMCResp structure is never actually set or
used.  So instead, it only ever wraps an error, or not.  This is much
more complicated than simply returning an error and makes tests and
readability of the code much worse.  This CR simply removes the VMCResp
structure in favor of returning an error.

Change-Id: I12db8441492eb8ad5bf4b52e2eea5f6bbd0ed4fd
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>