[FAB-9756] Add missing .env file
The e2e_cli test was recently updated based on the assumption that the
project name was e2e (as newer versions of docker compose seem to handle
directories with underscores in their names differently).  However, this
CR failed to check in a new file (.env) which set this.  This CR
remedies the situation.

Change-Id: I04967011b33789cb51d81c3e7738ade6554d7962
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>