FAB-9806 Remove GetVMName filter func
The interface for container.VM has a 'GetVMName(string, func(string)
(string, error)) (string, error)' definition.  However, no callers
outside of the dockercontroller package passes in a value for the
function of anything other than 'nil'.  This internal implementation
detail of dockercontroller does not need to be exposed outside of its
package, so this CR removes it.

Change-Id: I96351cf1559c8366e9b11a78466d497881acd395
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>