[FAB-9817] Use Tx-simulator after obtaining results
In the current implementation, a transaction simulator instance
releases itself (as an caution for optimization) when the caller
has obtained the simulation results. The assumption was that this
will be the last step in using the simulator.

However, now there is a situation where gossip would need to obtain the simulation
results and based on the private data present in the simulation results, it will use
the simulator to fetch the collection configurations.

This CR removes this self-release behavior. Also, a test is added to
make sure that the previously obtained rwset are not affected by
further data operations on the simulator.

Change-Id: Ic5664cd8d4536b31c603154ede573bbf3c095e76
Signed-off-by: manish <manish.sethi@gmail.com>